# Server/common/api/openapi.yaml
openapi: 3.1.1
info:
  title: Сайт с анекдотами Anekdotus
  version: 1.0.0
servers:
- url: /
paths:
  /api/v1/feed:
    get:
      operationId: apiV1FeedGET
      parameters:
      - explode: true
        in: query
        name: page
        required: false
        schema:
          default: 1
          type: integer
        style: form
      - explode: true
        in: query
        name: limit
        required: false
        schema:
          default: 10
          type: integer
        style: form
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AnekdotList"
          description: Успешный ответ с анекдотами
        "400":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
          description: Некоррекнтые параметры пагинации
        "503":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
          description: Разорвано соединение с БД
      summary: Получение списка анекдотов
      tags:
      - Feed
      x-eov-operation-handler: controllers/FeedController
  /api/v1/favourites:
    get:
      operationId: apiV1FavouritesGET
      parameters:
      - explode: true
        in: query
        name: page
        required: false
        schema:
          default: 1
          type: integer
        style: form
      - explode: true
        in: query
        name: limit
        required: false
        schema:
          default: 10
          type: integer
        style: form
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AnekdotList"
          description: Успешный ответ со списком избранных анекдотов
        "400":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
          description: Некоррекнтые параметры пагинации
        "401":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
          description: Не авторизован
        "403":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
          description: Нет прав
        "503":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
          description: Разорвано соединение с БД
      security:
      - bearerAuth: []
      summary: Получение списка избранных анекдотов
      tags:
      - Favourites
      x-eov-operation-handler: controllers/FavouritesController
    post:
      operationId: apiV1FavouritesPOST
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AddToFavouritesRequest"
        required: true
      responses:
        "201":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Anekdot"
          description: Анекдот успешно добавлен
        "401":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
          description: Не авторизован
        "403":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
          description: Нет прав
        "404":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
          description: Анекдот не найден
        "409":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
          description: Анекдот уже находится в избранном
        "503":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
          description: Разорвано соединение с БД
      security:
      - bearerAuth: []
      summary: Добавление анекдота в избранное по id
      tags:
      - Favourites
      x-eov-operation-handler: controllers/FavouritesController
  /api/v1/favourites/{id}:
    delete:
      operationId: apiV1FavouritesIdDELETE
      parameters:
      - description: Идентификатор анекдота
        explode: false
        in: path
        name: id
        required: true
        schema:
          type: integer
        style: simple
      responses:
        "204":
          description: Анекдот успешно удалён из избранного
        "401":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
          description: Не авторизован
        "403":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
          description: Нет прав
        "404":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
          description: Анекдот не найден
        "503":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
          description: Разорвано соединение с БД
      security:
      - bearerAuth: []
      summary: Удаление анекдота из избранного по id
      tags:
      - Favourites
      x-eov-operation-handler: controllers/FavouritesController
  /api/v1/register:
    post:
      operationId: apiV1RegisterPOST
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RegistrationRequest"
        required: true
      responses:
        "201":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Person"
          description: Успешная регистрация
        "400":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
          description: "Некорректный формат логина, пароля, имени или роли"
        "409":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
          description: Логин или имя заняты
        "503":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
          description: Разорвано соединение с БД
      summary: Регистрация пользователя
      tags:
      - Auth
      x-eov-operation-handler: controllers/AuthController
  /api/v1/login:
    post:
      operationId: apiV1LoginPOST
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginRequest"
        required: true
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Person"
          description: Успешный вход
        "400":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
          description: "Некорректный формат логина, пароля, имени или роли"
        "401":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
          description: Неверный логин или пароль
        "503":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
          description: Разорвано соединение с БД
      summary: Вход в аккаунт
      tags:
      - Auth
      x-eov-operation-handler: controllers/AuthController
  /api/v1/anekdots:
    post:
      operationId: apiV1AnekdotsPOST
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateAnekdotRequest"
        required: true
      responses:
        "201":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Anekdot"
          description: Анекдот успешно создан
        "401":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
          description: Не авторизован
        "403":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
          description: Нет прав
        "503":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
          description: Разорвано соединение с БД
      security:
      - bearerAuth: []
      summary: Загрузка нового анекдота
      tags:
      - Anekdot
      x-eov-operation-handler: controllers/AnekdotController
  /api/v1/anekdots/{id}:
    delete:
      operationId: apiV1AnekdotsIdDELETE
      parameters:
      - description: Идентификатор анекдота
        explode: false
        in: path
        name: id
        required: true
        schema:
          type: integer
        style: simple
      responses:
        "204":
          description: Анекдот успешно удален
        "400":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
          description: Неверный идентификатор
        "401":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
          description: Не авторизован
        "403":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
          description: Нет прав
        "404":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
          description: Анекдот не найден
        "503":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
          description: Разорвано соединение с БД
      security:
      - bearerAuth: []
      summary: Удаление анекдота по id
      tags:
      - Anekdot
      x-eov-operation-handler: controllers/AnekdotController
    put:
      operationId: apiV1AnekdotsIdPUT
      parameters:
      - description: Идентификатор анекдота
        explode: false
        in: path
        name: id
        required: true
        schema:
          example: 123
          minimum: 1
          type: integer
        style: simple
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateAnekdotRequest"
        required: true
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Anekdot"
          description: Анекдот успешно обновлен
        "400":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
          description: Неверные данные
        "401":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
          description: Не авторизован
        "403":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
          description: Нет прав
        "404":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
          description: Анекдот не найден
        "503":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
          description: Разорвано соединение с БД
      security:
      - bearerAuth: []
      summary: Редактирование анекдота по id
      tags:
      - Anekdot
      x-eov-operation-handler: controllers/AnekdotController
components:
  schemas:
    Anekdot:
      example:
        lastModifiedDate: 2025-10-07 12:58:05.000000
        hasBadWords: false
        id: 1
        text: Некая шутка
      properties:
        id:
          type: integer
        text:
          type: string
        hasBadWords:
          default: false
          description: Наличие в тексте анекдота ненормативной лексики
          type: boolean
        lastModifiedDate:
          type: string
      required:
      - hasBadWords
      - id
      - lastModifiedDate
      - text
    AnekdotList:
      items:
        $ref: "#/components/schemas/Anekdot"
      type: array
    Error:
      example:
        code: 5
        message: Логин занят
      properties:
        code:
          example: 5
          type: integer
        message:
          example: Логин занят
          type: string
      required:
      - code
      - message
    UserRole:
      description: |
        Роль пользователя:
        - 0 - USER
        - 1 - ADMIN
      enum:
      - 0
      - 1
      example: 0
      type: integer
    Person:
      example:
        role: 0
        name: name
        token: token
      properties:
        token:
          description: JWT токен для авторизации
          type: string
        name:
          type: string
        role:
          $ref: "#/components/schemas/UserRole"
      required:
      - name
      - role
      - token
    RegistrationRequest:
      example:
        password: secret123
        role: 0
        name: somename
        login: user123
      properties:
        login:
          example: user123
          maxLength: 10
          type: string
        password:
          example: secret123
          maxLength: 10
          type: string
        name:
          example: somename
          maxLength: 10
          type: string
        role:
          $ref: "#/components/schemas/UserRole"
      required:
      - login
      - name
      - password
      - role
    LoginRequest:
      example:
        password: secret123
        login: user123
      properties:
        login:
          example: user123
          maxLength: 10
          type: string
        password:
          example: secret123
          maxLength: 10
          type: string
      required:
      - login
      - password
    AddToFavouritesRequest:
      example:
        anekdot_id: 123
      properties:
        anekdot_id:
          example: 123
          maxLength: 10
          type: number
      required:
      - anekdot_id
    CreateAnekdotRequest:
      example:
        text: text
      properties:
        text:
          description: Текст анекдота или URL для парсинга
          minLength: 1
          type: string
      required:
      - text
    UpdateAnekdotRequest:
      example:
        text: Некий обновлённый текст анекдота
      properties:
        text:
          description: Новый текст анекдота
          example: Некий обновлённый текст анекдота
          minLength: 1
          type: string
      required:
      - text
  securitySchemes:
    bearerAuth:
      bearerFormat: JWT
      description: JWT токен полученный при входе или регистрации
      scheme: bearer
      type: http